Added 2 Enemy Types; Set Stats for single tier 2 towers

+++ Assets/Scripts/Tower/TowerInfo.cs
@@ -10,7 +10,7 @@ public class TowerInfo
 	private int cost;
 	private int damage;
 	private float attackSpeed;
-	private int range;
+	private float range;
 	private bool hasAOE;
 	private Dictionary<TowerType, TowerType> upgrades;
 
@@ -20,11 +20,11 @@ public class TowerInfo
 	public int Cost { get { return cost; } }
 	public int Damage { get { return damage; } }
 	public float AttackSpeed { get { return attackSpeed; } }
-	public int Range { get { return range; } }
+	public float Range { get { return range; } }
 	public bool AOE { get { return hasAOE; } }
 	public Dictionary<TowerType, TowerType> Upgrades { get { return upgrades; } }
 
-	public TowerInfo(Sprite towerSprite, Sprite bulletSprite, int cost, int damage, float attackSpeed, int range, bool hasAOE)
+	public TowerInfo(Sprite towerSprite, Sprite bulletSprite, int cost, int damage, float attackSpeed, float range, bool hasAOE)
 	{
 		this.towerSprite = towerSprite;
 		this.bulletSprite = bulletSprite;
@@ -36,6 +36,11 @@ public class TowerInfo
 		upgrades = new Dictionary<TowerType, TowerType>();
 	}
 
+	/// <summary>
+	/// Add a possible upgrade to this tower type
+	/// </summary>
+	/// <param name="secondaryType">The other tower type needed to upgrade</param>
+	/// <param name="upgradeType">The upgrade tower type result</param>
 	public void AddUpgrade(TowerType secondaryType, TowerType upgradeType)
 	{
 		upgrades.Add(secondaryType, upgradeType);
